# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Inazuma API",
        "description": "API for Inazuma Eleven",
        "contact": {
            "email": "gomez.sergio.smr@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/v1/teams": {
            "get": {
                "tags": [
                    "Teams"
                ],
                "summary": "List all teams with optional filters",
                "description": "Returns paginated list of teams. Different results are returned based on user role (admin, authenticated user, or guest).",
                "operationId": "66de14dea23a1cf63a1dee58a5cc5905",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "name[eq]",
                        "in": "query",
                        "description": "Filter by exact team name",
                        "schema": {
                            "type": "string",
                            "example": "Raimon"
                        }
                    },
                    {
                        "name": "userId[eq]",
                        "in": "query",
                        "description": "Filter by exact user ID who owns the team",
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "playerName[eq]",
                        "in": "query",
                        "description": "Filter teams containing players with exact name",
                        "schema": {
                            "type": "string",
                            "example": "Mark"
                        }
                    },
                    {
                        "name": "playerFullName[eq]",
                        "in": "query",
                        "description": "Filter teams containing players with exact full name",
                        "schema": {
                            "type": "string",
                            "example": "Mark Evans"
                        }
                    },
                    {
                        "name": "includePlayers",
                        "in": "query",
                        "description": "Include players in the response",
                        "schema": {
                            "type": "boolean",
                            "example": true
                        }
                    },
                    {
                        "name": "includeStats",
                        "in": "query",
                        "description": "Include player stats (requires includePlayers)",
                        "schema": {
                            "type": "boolean",
                            "example": true
                        }
                    },
                    {
                        "name": "includeTechniques",
                        "in": "query",
                        "description": "Include player techniques (requires includePlayers)",
                        "schema": {
                            "type": "boolean",
                            "example": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TeamCollection"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Teams"
                ],
                "summary": "Create a new team",
                "description": "Creates a new team with the specified players and their positions",
                "operationId": "fdf48d4cde576b3ff280b0e588981c4f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TeamRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Team created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TeamResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/teams/{id}": {
            "get": {
                "tags": [
                    "Teams"
                ],
                "summary": "Get team details",
                "description": "Returns detailed information about a specific team",
                "operationId": "bf2640af970d006d2afac946fd93f450",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Team ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "includePlayers",
                        "in": "query",
                        "description": "Include players in the response",
                        "schema": {
                            "type": "boolean",
                            "example": true
                        }
                    },
                    {
                        "name": "includeStats",
                        "in": "query",
                        "description": "Include player stats (requires includePlayers)",
                        "schema": {
                            "type": "boolean",
                            "example": true
                        }
                    },
                    {
                        "name": "includeTechniques",
                        "in": "query",
                        "description": "Include player techniques (requires includePlayers)",
                        "schema": {
                            "type": "boolean",
                            "example": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Team details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TeamResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - not authorized to view this team"
                    },
                    "404": {
                        "description": "Team not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Teams"
                ],
                "summary": "Update team information",
                "description": "Updates team details and player positions",
                "operationId": "4b6c0c9fc6aaa568cbd405d57a878fbf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Team ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TeamRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Team updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TeamResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - not authorized to update this team"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Teams"
                ],
                "summary": "Delete a team",
                "description": "Permanently deletes a team",
                "operationId": "11e231e8473c3df157fc2660394a1ba1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Team ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Team deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Team deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - not authorized to delete this team"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Coach": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Seymour Hillman"
                    },
                    "image": {
                        "type": "string",
                        "example": "https://static.wikia.nocookie.net/inazuma/images/b/b8/Hillman_%28Sprite%29.png/revision/latest?cb=20201025114237&path-prefix=es"
                    },
                    "version": {
                        "type": "string",
                        "example": "ie1"
                    }
                },
                "type": "object"
            },
...
            "Formation": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Diamante"
                    },
                    "layout": {
                        "type": "string",
                        "example": "4-3-3"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "TeamPosition": {
                "type": "string",
                "enum": [
                    "pos-0",
                    "pos-1",
                    "pos-2",
                    "pos-3",
                    "pos-4",
                    "pos-5",
                    "pos-6",
                    "pos-7",
                    "pos-8",
                    "pos-9",
                    "pos-10",
                    "bench-0",
                    "bench-1",
                    "bench-2",
                    "bench-3",
                    "bench-4"
                ],
                "example": "pos-1"
            },
...
            "TeamRequest": {
                "required": [
                    "name",
                    "formation_id",
                    "emblem_id",
                    "coach_id",
                    "players"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Raimon"
                    },
                    "formation_id": {
                        "description": "ID of the formation",
                        "type": "integer",
                        "example": 1
                    },
                    "emblem_id": {
                        "description": "ID of the emblem",
                        "type": "integer",
                        "example": 1
                    },
                    "coach_id": {
                        "description": "ID of the coach",
                        "type": "integer",
                        "example": 1
                    },
                    "players": {
                        "type": "array",
                        "items": {
                            "required": [
                                "player_id",
                                "position"
                            ],
                            "properties": {
                                "player_id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "position": {
                                    "$ref": "#/components/schemas/TeamPosition"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "TeamResource": {
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/Team"
                    }
                },
                "type": "object"
            },
            "Team": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Raimon"
                    },
                    "formation_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "formation": {
                        "$ref": "#/components/schemas/Formation"
                    },
                    "emblem_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "emblem": {
                        "$ref": "#/components/schemas/Emblem"
                    },
                    "coach_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "coach": {
                        "$ref": "#/components/schemas/Coach"
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "players": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "name": {
                                    "type": "string",
                                    "example": "Mark"
                                },
                                "full_name": {
                                    "type": "string",
                                    "example": "Mark Evans"
                                },
                                "position": {
                                    "type": "string",
                                    "example": "Portero"
                                },
                                "pivot": {
                                    "$ref": "#/components/schemas/TeamPlayerPivot"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "TeamCollection": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Team"
                        }
                    },
                    "links": {
                        "properties": {
                            "first": {
                                "type": "string",
                                "example": "http://localhost:8000/api/v1/teams?page=1"
                            },
                            "last": {
                                "type": "string",
                                "example": "http://localhost:8000/api/v1/teams?page=3"
                            },
                            "prev": {
                                "type": "string",
                                "nullable": true
                            },
                            "next": {
                                "type": "string",
                                "example": "http://localhost:8000/api/v1/teams?page=2"
                            }
                        },
                        "type": "object"
                    },
                    "meta": {
                        "properties": {
                            "current_page": {
                                "type": "integer",
                                "example": 1
                            },
                            "last_page": {
                                "type": "integer",
                                "example": 3
                            },
                            "per_page": {
                                "type": "integer",
                                "example": 15
                            },
                            "total": {
                                "type": "integer",
                                "example": 35
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}